changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11757 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Misc Objects/DataStructures/ORSafeBuffer.h
@@ -29,10 +29,10 @@
 	NSLock* bufferLock;
 }
 - (id) init;
-- (id) initWithBufferSize:(unsigned)aSize;
+- (id) initWithBufferSize:(NSUInteger)aSize;
 - (void) dealloc;
-- (id) objectAtIndex:(unsigned)index;
+- (id) objectAtIndex:(NSUInteger)index;
 - (id) lastObject;
 - (void) addObject:(id) anObject;
-- (unsigned) count;
+- (NSUInteger) count;
 @end
